pyproject.toml
[project]
name = "alfred-gitmoji"
version = "0.1.0"
description = "Search for gitmoji and copy them to the clipboard easily with alfred"
authors = [
    {name = "Eugene Jeon",email = "eugenejeon.me@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "pyinstaller (>=6.12.0,<7.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
